backend/pyproject.toml
[project]
name = "stepthrough-backend"
version = "0.1.0"
description = "Local-first backend for walkthrough screenshot extraction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "pillow>=10.4.0",
  "pydantic>=2.9.0",
  "python-multipart>=0.0.12",
  "scenedetect[opencv]>=0.6.4",
  "uvicorn[standard]>=0.32.0"
]

[dependency-groups]
dev = [
  "httpx>=0.28.0",
  "pytest>=8.3.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
